refactor: isolate native CLI availability resolver
